Remove react-dom-polyfill
When I required this package in a project it complained about a missing
package (react-dom-polyfill).
Instead of adding react-dom-polyfill as a dependency to this package I
just removed it.
So that it's possible to require this package without adding the extra
react-dom-polyfill as a dependency.
I think everyone has had enough time to transition to react > 0.14 :)